refactor: seperate view and controller in Add
+ ProductAdd.js
Controller

+ ProductAddTemplate.js
View
